Cambios en el controlador del backend, ahora al hacer delete, también elimina la imagen asociada al artículo.

+++ Backend/controllers/article.js
@@ -46,20 +46,20 @@ var controller = {
         var params = request.body;
 
         //  2.- Validar datos con Validator
-        try{
+        try {
 
             var validate_title = !validator.isEmpty(params.title);
             var validate_content = !validator.isEmpty(params.content);
 
         }
-        catch(err){
+        catch (err) {
             return response.status(200).send({
                 status: 'error',
                 message: 'Faltan datos por enviar.'
             });
         }
 
-        if(validate_title && validate_content){
+        if (validate_title && validate_content) {
 
             //  3.- Crear el objeto a guardar
 
@@ -69,13 +69,24 @@ var controller = {
 
             article.title = params.title;
             article.content = params.content;
-            article.image = null;
 
+            if (params.image) {
+
+                article.image = params.image;
+
+            } else {
+
+                article.image = null;
+
+            }
+            console.log('BACKEND');
+            console.log(article);
+            console.log('FIN BACKEND');
             //  5.- Guardar el articulo
-            
+
             article.save((err, articleStored) => {
 
-                if( err || !articleStored ){
+                if (err || !articleStored) {
 
                     return response.status(404).send({
                         status: 'error',
@@ -92,8 +103,8 @@ var controller = {
 
             });
 
-            
-        }else{
+
+        } else {
 
             return response.status(200).send({
                 status: 'error',
@@ -107,20 +118,20 @@ var controller = {
 
         //Sacar ultimos articulos
 
-        var query = Article.find({  });             // Variable con la query
+        var query = Article.find({});             // Variable con la query
 
         var limit = request.params.limit;           //Recogemos el parametro limit de la url
-        
-        if( limit || limit != undefined ){          //Comprobamos que exista y no sea undefined
+
+        if (limit || limit != undefined) {          //Comprobamos que exista y no sea undefined
 
             query.limit(parseInt(limit));           //Aplicamos el valor que le pasamos por la url para limitar los contenidos
 
         }
-        
+
         //  FIND (-_id ordena por id en orden inverso, más nuevos primero)
         query.sort('-_id').exec((err, articles) => {
 
-            if(err){
+            if (err) {
 
                 return response.status(500).send({
                     status: 'error',
@@ -129,7 +140,7 @@ var controller = {
 
             }
 
-            if(!articles){
+            if (!articles) {
 
                 return response.status(404).send({
                     status: 'error',
@@ -145,7 +156,7 @@ var controller = {
 
         });
 
-        
+
     },
     getArticle: (request, response) => {
 
@@ -153,7 +164,7 @@ var controller = {
         var articleId = request.params.id;
 
         //  Comprobar que existe
-        if( !articleId || articleId == null ){
+        if (!articleId || articleId == null) {
 
             return response.status(404).send({
                 status: 'error',
@@ -165,7 +176,7 @@ var controller = {
         //  Buscar Artículo
         Article.findById(articleId, (err, article) => {
 
-            if(err || !article){
+            if (err || !article) {
                 return response.status(404).send({
                     status: 'error',
                     message: 'No existe el artículo.'
@@ -192,17 +203,17 @@ var controller = {
         var params = request.body;
 
         //Validar los datos
-        try{
+        try {
 
             var validate_title = !validator.isEmpty(params.title);
 
             var validate_content = !validator.isEmpty(params.content);
 
-            
+
 
             //Enviar una respuesta
 
-        }catch(err){
+        } catch (err) {
 
             return response.status(200).send({
                 status: 'error',
@@ -211,20 +222,20 @@ var controller = {
 
         }
 
-        if(validate_title && validate_content){
+        if (validate_title && validate_content) {
 
             //Hacer un find and update
             //new: true nos duelve un JSON con el elemto actualizado
             Article.findOneAndUpdate({ _id: articleId }, params, { new: true }, (err, articleUpdated) => {
 
-                if(err){
+                if (err) {
                     return response.status(500).send({
                         status: 'error',
                         message: 'Error al actualizar.'
                     });
                 }
 
-                if(!articleUpdated){
+                if (!articleUpdated) {
                     return response.status(404).send({
                         status: 'error',
                         message: 'No existe el articulo a modificar.'
@@ -238,7 +249,7 @@ var controller = {
 
             });
 
-        }else{
+        } else {
             return response.status(200).send({
                 status: 'error',
                 message: 'La validación no es correcta.'
@@ -246,34 +257,64 @@ var controller = {
         }
 
     },
-    delete: (request, response) => {
+    delete: (req, res) => {
+        // Recogemos el id del articulo por la url
+        var articleId = req.params.id;
 
-        //Recoger el ID de la URL
-        var articleId = request.params.id;
-        //Hacer un find and delete
-        Article.findOneAndDelete({ _id: articleId }, (err, articleRemoved) => {
 
-            if(err){
-                return response.status(500).send({
+        // Comprobar que si existe
+        if (!articleId || articleId == null || articleId == undefined) {
+            return res.status(404).send({
+                status: 'error',
+                message: 'No existe el articulo!!!'
+            })
+        }
+
+        // Buscar el articulo
+        Article.findById(articleId, (err, article) => {
+            if (err || !article) {
+                return res.status(404).send({
                     status: 'error',
-                    message: 'Error al eliminar.'
+                    message: 'No existe el articulo!!!'
+                })
+            }
+
+            if (article.image) {
+                var file_name = article.image;
+                var file_path = 'upload/articles/' + file_name;
+
+                //console.log(file_path)
+
+                fs.unlink(file_path, function (err) {
+                    if (err) throw err;
+                    console.log('file deleted');
                 });
             }
 
-            if(!articleRemoved){
-                return response.status(404).send({
-                    status: 'error',
-                    message: 'No existe el articulo a eliminar.'
-                });
-            }
 
-            return response.status(200).send({
-                status: 'success',
-                article: articleRemoved
-            });
 
-        });
+            // find and Delete
+            Article.findOneAndDelete({ _id: articleId }, (err, articleRemoved) => {
+                if (err) {
+                    return res.status(500).send({
+                        status: 'error',
+                        message: 'Error al eliminar!!!'
+                    })
+                }
 
+                if (!articleRemoved) {
+                    return res.status(404).send({
+                        status: 'error',
+                        message: 'No se ha borrado el articulo, posiblemente no exista!!!'
+                    })
+                }
+
+                return res.status(200).send({
+                    status: 'success',
+                    article: articleRemoved
+                })
+            })
+        })
     },
     upload: (request, response) => {
 
@@ -281,8 +322,8 @@ var controller = {
 
         //Recoger el fichero de la petición
         var file_name = 'Imagen no subida...';
-        
-        if(!request.files){
+
+        if (!request.files) {
             return response.status(404).send({
                 status: 'error',
                 message: file_name
@@ -304,46 +345,55 @@ var controller = {
         var file_extension = file_name_split[1];
 
         //Comprobar la extensión, sólo imágenes, si no es válida la extensión, borrar el fichero
-        if(file_extension != 'jpg' && file_extension != 'jpeg' && file_extension != 'png' && file_extension != 'gif'){
-            
+        if (file_extension != 'jpg' && file_extension != 'jpeg' && file_extension != 'png' && file_extension != 'gif') {
+
             //Borrar el archivo subido
             fs.unlink(file_path, (err) => {
-                
+
                 return response.status(500).send({
                     status: 'error',
                     message: 'La extensión de la imágen no es válida.'
                 });
-                
-            });
-            
 
-        }else{
+            });
+
+
+        } else {
             //Si todo es válido, sacamos ID de la URL
             var articleId = request.params.id;
-            //Buscar el articulo, asignarle el nombre de la imagen y actualizarlo
-            Article.findOneAndUpdate({ _id: articleId }, { image: file_name }, { new: true }, (err, articleUpdated) => {
 
-                if(err){
-                    return response.status(500).send({
-                        status: 'error',
-                        message: 'Error al agregar/modificar la imagen.'
+            if (articleId) {
+                //Buscar el articulo, asignarle el nombre de la imagen y actualizarlo
+                Article.findOneAndUpdate({ _id: articleId }, { image: file_name }, { new: true }, (err, articleUpdated) => {
+
+                    if (err) {
+                        return response.status(500).send({
+                            status: 'error',
+                            message: 'Error al agregar/modificar la imagen.'
+                        });
+                    }
+
+                    if (!articleUpdated) {
+                        return response.status(404).send({
+                            status: 'error',
+                            message: 'No existe el articulo al cual quiere agregar/modificar la imagen.'
+                        });
+                    }
+
+                    return response.status(200).send({
+                        status: 'success',
+                        article: articleUpdated
                     });
-                }
-
-                if(!articleUpdated){
-                    return response.status(404).send({
-                        status: 'error',
-                        message: 'No existe el articulo al cual quiere agregar/modificar la imagen.'
-                    });
-                }
-
+                });
+            } else {
                 return response.status(200).send({
                     status: 'success',
-                    article: articleUpdated
+                    image: file_name
                 });
-            });
+            }
+
         }
-        
+
     },
     getImage: (request, response) => {
 
@@ -354,11 +404,11 @@ var controller = {
         var path_file = './upload/articles/' + file;
 
         //Comprobamos que la ruta existe
-        if(fs.existsSync(path_file)){
+        if (fs.existsSync(path_file)) {
 
             return response.sendFile(path.resolve(path_file));
 
-        }else{
+        } else {
             return response.status(404).send({
                 status: 'error',
                 message: 'La imagen no existe.'
@@ -384,29 +434,29 @@ var controller = {
                 { 'content': { "$regex": searchString, "$options": "i" } }
             ]
         })
-        .sort([[ 'date', 'descending' ]])
-        .exec( (err, articles) => {
+            .sort([['date', 'descending']])
+            .exec((err, articles) => {
 
-            if(err){
-                return response.status(500).send({
-                    status: 'error',
-                    message: 'Error al buscar los artículos.'
+                if (err) {
+                    return response.status(500).send({
+                        status: 'error',
+                        message: 'Error al buscar los artículos.'
+                    });
+                }
+
+                if (!articles || articles.length <= 0) {
+                    return response.status(404).send({
+                        status: 'error',
+                        message: 'No existen articulos que coincidan con la búsqueda.'
+                    });
+                }
+
+                return response.status(200).send({
+                    status: 'success',
+                    articles
                 });
-            }
 
-            if(!articles || articles.length <= 0){
-                return response.status(404).send({
-                    status: 'error',
-                    message: 'No existen articulos que coincidan con la búsqueda.'
-                });
-            }
-
-            return response.status(200).send({
-                status: 'success',
-                articles
-            });
-
-        });//FIN DEL EXEC DEL FIND
+            });//FIN DEL EXEC DEL FIND
     }
 
 }; //END CONTROLLER
